gridflag : intermediate commit
The graph building works and runs through from end to end, but the function
calls still yield the wrong results. The intention was to generate a histogram
per uv-pixel and use that histogram to calculate the median and MAD per pixel.
However at the moment it simply accumulates all the points per pixel before
calculation which becomes untenable for large datasets.