Hogbom clean with scikit-build-core
Cleaned up original fortran to C++ translated implementation, templated
pybind11 bindings to allow either floats or doubles.

Included both a clean version with logging callbacks and one without, depending
on which way we want to go with the Python code.

pyproject.toml now calls CMake to build the extension. Template for a time when
we need to wrap cuda/Kokkos.