Switched to templated C++ code
The earlier bindings were converting from double --> float if an array
of doubles was passed in. The new templating should ensure that doubles
are treated as doubles the entire way through.

This should help mitigate issues with deconvolving images that have
signal close to the floating point noise floor, and with added numerical
noise going from double --> float --> double.