Add Jupyter notebook with demo
Added Jupyter notebook with a demo comparing the astroviper hogbom clean
implementation to the existing CASA implementation.

Found a small bug in the deconvolution.py hogbom function, fixed that as
well. The issue was that because of the templated nature of the pybind
wrapper, the arguments must be provided in recognizable types. I was
passing in a list where it expected a numpy array, and that broke
something internally. It did not perform the conversion automatically,
like one would expect from a generic numpy function etc.