Fix deconv param checking, drop redundant params
Fixed the deconvolution parameter validation, and removed the model and
redisual arrays (not xds) from being returned. The raw data arrays for a
single plane were also being returned. This has now been fixed to remove
the raw arrays, accumulate the arrays into the xds (already happening
before) and return the xds (new in this commit).