Fix ipynb to use toolviper data + bugfixes
Fixed the ipynb to use the toolviper data repo, which dropped the
requirement for casatasks. The casatasks import was making the test
runners fall over.

I ran a local copilot code review and fixed a couple of issues that it
flagged - these did not seem to be active bugs, but more "best
practices".