Add new ReturnDict class (broken commit)
Experimenting with implementing a ReturnDict class, as a workaround for
using nested dictionaries.

The ReturnDict class is essentially a convenience wrapper around a
regular nested dict. The dictionary is indexed via a Tuple (specifically
a NamedTuple) allowing for a unique combination of time/freq/pol
selections to index across any keys in O(N).

It also has convenience functions to return all keys with a partial
match, say chan=3 or pol=0.

The underlying structure is :

self.data = {(time=0,pol=0,chan=0):{...},
             (time=1,pol=0,chan=0):{...},
             ...
            }

The ReturnDict class is fully implemented, but it has not been put into
action within the deconvolution function.