Add max PSF sidelobe to deconvolution summaryminor
Merged the latest changes from 45-fit_gaussian, modified it to return
the max PSF sidelobe. This is now returned in the return dictionary from
the deconvolver.

The stopcode is not yet populated, this is intentional. Once the
stopcodes are finalized and documented for iteration control, it will be
straightforward to add it to the deconvolver.