Bugfixes + expand test coverage
Fixed a handful of bugs in the deconvolve() function, related to moving
the per-plane iteration outside of hogbom_clean and into deconvolve()
where it it more appropriate.

Added tests for the deconvolve() function, but at the moment they are
failing.

The hogbom clean function and python bindings need to be refactored to
accept only 2D arrays, and move all the cube slicing to be outside the
actual deconvolution part. This will be done in a subsequent commit.