Initial refactor of hclean.cpp
The hogbom clean C++ implementation and associated Python bindings have
been refactored to remove the iteration over polarization. This is the
initial version of the refactor, and has not been tested for correctness
etc.

The deconvolve() function will do all the iteration, and the lower level
deconvolution functions will only accept a single plane to run the
deconvolution on.