Fleshed out iteration control class
Filled in the functions for iteration control, adapted from _gclean.py
within the CASA codebase.

The function has been refactored to use the new ReturnDict structure,
which is flat compared with the nested return dictionary in CASA.

The implementation has been tested in isolation, but not in conjuction
with the gridder and deconvolver, which will be done subsequently.