Fix iteration_control notebook, add unit tests
Added the unit tests for the iteration-control module. Made some small fixes to
the notebook, since I moved the path of iteration_control to live in
core/imaging/imaging_utils rather than just core/imaging/

Added a convenience function in the ReturnDict to reset the StopCodes to zero,
while keeping the rest of the dict intact.