Updates to return_dict behaviour
The return dictionary now correctly supports lists (previously would
over-write, rather than append) and fixed the merging logic in the iteration
control module to account for this.

This allows for tracking of convergence parameters to generate a history plot,
and print out useful stats to the user at the end.